Move permissions to a separate file
