+++ apps/protocol-api/src/main.ts
@@ -9,25 +9,15 @@ import { LinearClient } from '@linear/sdk';
 
 import { resolvers } from './prisma/generated/type-graphql';
 import { customResolvers } from './prisma/resolvers';
-import { deny, or, rule, shield } from 'graphql-shield';
 import { graphqlHTTP } from 'express-graphql';
 import fetch from 'cross-fetch';
+import { permissions } from './permissions';
 
 import cors = require('cors');
 
 console.log('Starting');
 const prisma = new PrismaClient();
-const AIRTABLE_API_TOKEN = process.env.AIRTABlE_API_TOKEN;
-const KEVIN_MALONE_TOKEN = process.env.KEVIN_MALONE_TOKEN;
-const LINEAR_JOB_TOKEN = process.env.LINEAR_JOB_TOKEN;
-const CONTRACT_SYNC_JOB_TOKEN = process.env.CONTRACT_SYNC_JOB_TOKEN;
 
-const BACKEND_TOKENS = [
-  AIRTABLE_API_TOKEN,
-  KEVIN_MALONE_TOKEN,
-  LINEAR_JOB_TOKEN,
-  CONTRACT_SYNC_JOB_TOKEN,
-];
 const LINEAR_TOKEN_URL = 'https://api.linear.app/oauth/token';
 const LINEAR_REDIRECT_URI = process.env.LINEAR_REDIRECT_URI;
 const LINEAR_CLIENT_ID = process.env.LINEAR_CLIENT_ID;
@@ -37,346 +27,6 @@ const PROTOCOL_FRONTEND = process.env.PROTOCOL_FRONTEND;
 const typeSchema = buildSchemaSync({
   resolvers: [...resolvers, ...customResolvers],
 });
-
-const isAuthenticated = rule()(async (parent, args, ctx, info) => {
-  if (!ctx.req.session.siwe) {
-    return false;
-  }
-  return true;
-});
-
-const hasToken = rule()(async (parent, args, ctx, info) => {
-  const auth = ctx.req.headers['authorization'];
-  if (auth) {
-    const found = BACKEND_TOKENS.find(token => token === auth);
-    return !!found;
-  }
-  return false;
-});
-
-const permissions = shield(
-  {
-    Query: {
-      '*': deny,
-      chain: hasToken,
-      contribution: or(isAuthenticated, hasToken),
-      contributions: or(isAuthenticated, hasToken),
-      activityTypes: or(isAuthenticated, hasToken),
-      attestations: isAuthenticated,
-      listActivityTypesByUser: isAuthenticated,
-      getUser: isAuthenticated,
-      guild: or(isAuthenticated, hasToken),
-      guilds: or(isAuthenticated, hasToken),
-      guildUsers: or(isAuthenticated, hasToken),
-      guildImports: or(isAuthenticated, hasToken),
-      listUserByAddress: isAuthenticated,
-      getContributionCountByDateForUserInRange: or(isAuthenticated, hasToken),
-      getDaoContributionCountByUser: or(isAuthenticated, hasToken),
-      getDaoContributionCount: or(isAuthenticated, hasToken),
-      getContributionCountByActivityType: or(isAuthenticated, hasToken),
-      users: hasToken,
-      jobRuns: hasToken,
-      linearUsers: hasToken,
-      linearIssues: hasToken,
-      contributionStatuses: hasToken,
-      getActiveGuildUsersAverage: or(isAuthenticated, hasToken),
-    },
-    ContributionCountByUser: or(isAuthenticated, hasToken),
-    ContributionCountByDate: or(isAuthenticated, hasToken),
-    ContributionCountByActivityType: or(isAuthenticated, hasToken),
-    Mutation: {
-      '*': deny,
-      createOneActivityType: hasToken,
-      createOneContribution: hasToken,
-      createOneGuild: hasToken,
-      createOneGuildUser: or(isAuthenticated, hasToken),
-      createOneJobRun: hasToken,
-      createManyAttestation: hasToken,
-      createManyContribution: hasToken,
-      createManyLinearIssue: hasToken,
-      createOnChainUserContribution: isAuthenticated,
-      createOneUser: or(isAuthenticated, hasToken),
-      createOneUserActivity: hasToken,
-      createUserAttestation: isAuthenticated,
-      createUserContribution: isAuthenticated,
-      createUserCustom: or(hasToken, isAuthenticated),
-      createGuildUserCustom: isAuthenticated,
-      createUserOnChainAttestation: isAuthenticated,
-      deleteOneContribution: or(hasToken, isAuthenticated),
-      deleteUserContribution: or(isAuthenticated, hasToken),
-      getOrCreateActivityType: isAuthenticated,
-      updateOneContribution: hasToken,
-      updateOneGuild: hasToken,
-      updateOneUser: hasToken,
-      updateUserContribution: isAuthenticated,
-      updateGuildCustom: isAuthenticated,
-      updateUserCustom: isAuthenticated,
-      updateGuildUserCustom: isAuthenticated,
-      updateUserOnChainAttestation: isAuthenticated,
-      updateUserOnChainContribution: isAuthenticated,
-      upsertOneAttestation: hasToken,
-      upsertOneActivityType: hasToken,
-      upsertOneContribution: hasToken,
-      upsertOneLinearCycle: hasToken,
-      upsertOneLinearIssue: hasToken,
-      upsertOneLinearProject: hasToken,
-      upsertOneLinearTeam: hasToken,
-      upsertOneLinearUser: hasToken,
-    },
-    ActivityType: {
-      id: or(isAuthenticated, hasToken),
-      createdAt: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      name: or(isAuthenticated, hasToken),
-      active: or(isAuthenticated, hasToken),
-      default: or(isAuthenticated, hasToken),
-      users: or(isAuthenticated, hasToken),
-      contributions: or(isAuthenticated, hasToken),
-      categoryActivity: or(isAuthenticated, hasToken),
-      guilds: or(isAuthenticated, hasToken),
-    },
-    Attestation: {
-      id: or(isAuthenticated, hasToken),
-      confidence: or(isAuthenticated, hasToken),
-      contribution: or(isAuthenticated, hasToken),
-      user: or(isAuthenticated, hasToken),
-      createdAt: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      date_of_attestation: or(isAuthenticated, hasToken),
-      user_id: or(isAuthenticated, hasToken),
-      attestation_status: or(isAuthenticated, hasToken),
-    },
-    AttestationConfidence: {
-      name: or(isAuthenticated, hasToken),
-      createdAt: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      id: or(isAuthenticated, hasToken),
-    },
-    AttestationStatus: {
-      id: or(isAuthenticated, hasToken),
-      name: or(isAuthenticated, hasToken),
-    },
-    ChainType: {
-      id: or(isAuthenticated, hasToken),
-      createdAt: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      name: or(isAuthenticated, hasToken),
-      users: or(isAuthenticated, hasToken),
-    },
-    Chain: {
-      id: hasToken,
-      createdAt: hasToken,
-      updatedAt: hasToken,
-      name: hasToken,
-      chain_id: or(isAuthenticated, hasToken),
-    },
-    Contribution: {
-      id: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      name: or(isAuthenticated, hasToken),
-      status_id: or(isAuthenticated, hasToken),
-      status: or(isAuthenticated, hasToken),
-      activity_type_id: or(isAuthenticated, hasToken),
-      activity_type: or(isAuthenticated, hasToken),
-      user_id: or(isAuthenticated, hasToken),
-      user: or(isAuthenticated, hasToken),
-      date_of_submission: or(isAuthenticated, hasToken),
-      date_of_engagement: or(isAuthenticated, hasToken),
-      details: or(isAuthenticated, hasToken),
-      proof: or(isAuthenticated, hasToken),
-      attestations: or(isAuthenticated, hasToken),
-      partners: or(isAuthenticated, hasToken),
-      guilds: or(isAuthenticated, hasToken),
-      linear_issue: or(isAuthenticated, hasToken),
-      twitter_tweet_contributions: or(isAuthenticated, hasToken),
-      on_chain_id: or(isAuthenticated, hasToken),
-      tx_hash: or(isAuthenticated, hasToken),
-      chain: or(isAuthenticated, hasToken),
-    },
-    ContributionStatus: {
-      id: or(isAuthenticated, hasToken),
-      createdAt: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      name: or(isAuthenticated, hasToken),
-      contributions: or(isAuthenticated, hasToken),
-    },
-    DiscordUser: {
-      createdAt: or(isAuthenticated, hasToken),
-      discord_id: or(isAuthenticated, hasToken),
-      display_name: or(isAuthenticated, hasToken),
-      id: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      user: or(isAuthenticated, hasToken),
-      user_id: or(isAuthenticated, hasToken),
-    },
-    Guild: {
-      activity_type: or(isAuthenticated, hasToken),
-      congrats_channel: or(isAuthenticated, hasToken),
-      contributions: or(isAuthenticated, hasToken),
-      createdAt: or(isAuthenticated, hasToken),
-      discord_id: or(isAuthenticated, hasToken),
-      id: or(isAuthenticated, hasToken),
-      logo: or(isAuthenticated, hasToken),
-      name: or(isAuthenticated, hasToken),
-      twitter_accounts: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      users: or(isAuthenticated, hasToken),
-      contribution_reporting_channel: or(hasToken, isAuthenticated),
-      status: or(hasToken, isAuthenticated),
-    },
-    GuildContribution: {
-      id: or(isAuthenticated, hasToken),
-      contribution_id: or(isAuthenticated, hasToken),
-      guild_id: or(isAuthenticated, hasToken),
-      guild: or(isAuthenticated, hasToken),
-    },
-    GuildMembershipStatus: {
-      id: or(isAuthenticated, hasToken),
-      createdAt: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      name: or(isAuthenticated, hasToken),
-    },
-    GuildImport: {
-      id: or(isAuthenticated, hasToken),
-      createdAt: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      guild_id: or(isAuthenticated, hasToken),
-      guild: or(isAuthenticated, hasToken),
-      integration_type_id: or(isAuthenticated, hasToken),
-      integration_type: or(isAuthenticated, hasToken),
-      authentication_token: or(isAuthenticated, hasToken),
-      import_status: or(isAuthenticated, hasToken),
-      users: or(isAuthenticated, hasToken),
-    },
-    GuildImportStatus: {
-      id: or(isAuthenticated, hasToken),
-      createdAt: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      name: or(isAuthenticated, hasToken),
-    },
-    GuildUser: {
-      id: or(isAuthenticated, hasToken),
-      createdAt: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      user: or(isAuthenticated, hasToken),
-      user_id: or(isAuthenticated, hasToken),
-      guild: or(isAuthenticated, hasToken),
-      guild_id: or(isAuthenticated, hasToken),
-      favorite: or(isAuthenticated, hasToken),
-      membershipStatus: or(isAuthenticated, hasToken),
-      membership_status_id: or(isAuthenticated, hasToken),
-    },
-    IntegrationType: {
-      id: or(isAuthenticated, hasToken),
-      name: or(isAuthenticated, hasToken),
-    },
-    TwitterUser: {
-      id: or(isAuthenticated, hasToken),
-      name: or(isAuthenticated, hasToken),
-      username: or(isAuthenticated, hasToken),
-    },
-    AffectedRowsOutput: {
-      count: hasToken,
-    },
-    JobRun: {
-      id: hasToken,
-      createdAt: hasToken,
-      updatedAt: hasToken,
-      completedDate: hasToken,
-      startDate: hasToken,
-      name: hasToken,
-    },
-    User: {
-      id: or(isAuthenticated, hasToken),
-      createdAt: or(isAuthenticated, hasToken),
-      updatedAt: or(isAuthenticated, hasToken),
-      name: or(isAuthenticated, hasToken),
-      display_name: or(isAuthenticated, hasToken),
-      address: or(isAuthenticated, hasToken),
-      chain_type: or(isAuthenticated, hasToken),
-      full_name: or(isAuthenticated, hasToken),
-      discord_users: or(isAuthenticated, hasToken),
-      guild_users: or(isAuthenticated, hasToken),
-      twitter_user: or(isAuthenticated, hasToken),
-      active: or(isAuthenticated, hasToken),
-      linear_users: or(isAuthenticated, hasToken),
-    },
-    UserActivity: {
-      id: hasToken,
-      createdAt: hasToken,
-      updatedAt: hasToken,
-      activity_type: hasToken,
-      user: hasToken,
-    },
-    LinearIssue: {
-      id: hasToken,
-      archivedAt: hasToken,
-      assignee_id: hasToken,
-      autoArchivedAt: hasToken,
-      autoClosedAt: hasToken,
-      boardOrder: hasToken,
-      branchName: hasToken,
-      canceledAt: hasToken,
-      completedAt: hasToken,
-      createdAt: hasToken,
-      creator_id: hasToken,
-      customerTickerCount: hasToken,
-      cycle_id: hasToken,
-      description: hasToken,
-      dueDate: hasToken,
-      estimate: hasToken,
-      identifier: hasToken,
-      linear_id: hasToken,
-      priority: hasToken,
-      priorityLabel: hasToken,
-      project_id: hasToken,
-      snoozedUntilAt: hasToken,
-      sortOrder: hasToken,
-      startedAt: hasToken,
-      subIssueSortOrder: hasToken,
-      team_id: hasToken,
-      title: hasToken,
-      trashed: hasToken,
-      updatedAt: hasToken,
-      url: hasToken,
-    },
-    LinearTeam: {
-      linear_id: hasToken,
-      name: hasToken,
-      key: hasToken,
-      id: hasToken,
-    },
-    LinearProject: {
-      linear_id: hasToken,
-      name: hasToken,
-      id: hasToken,
-    },
-    LinearCycle: {
-      id: hasToken,
-      endsAt: hasToken,
-      linear_id: hasToken,
-      number: hasToken,
-      startsAt: hasToken,
-    },
-    LinearUser: {
-      id: or(isAuthenticated, hasToken),
-      active: hasToken,
-      displayName: hasToken,
-      email: hasToken,
-      linear_id: hasToken,
-      name: hasToken,
-      url: hasToken,
-      createdAt: hasToken,
-      access_token: hasToken,
-      active_token: or(isAuthenticated, hasToken),
-      user_id: hasToken,
-    },
-  },
-  {
-    fallbackRule: deny,
-    debug: process.env.NODE_ENV === 'development' ? true : false,
-  },
-);
 
 const schema = applyMiddleware(typeSchema, permissions);
