+++ apps/protocol-api/src/permisisons/inputs.ts
@@ -1,0 +1,2 @@
+// Create individual input rules
+// Create more generic user and admin rulesj
